Add RISC-V PLIC support

+++ ostd/src/irq/bottom_half.rs
@@ -18,7 +18,7 @@ use crate::task::disable_preempt;
 ///
 /// The function may be called only once; subsequent calls take no effect.
 pub fn register_bottom_half_handler_l1(
-    func: fn(DisabledLocalIrqGuard, usize) -> DisabledLocalIrqGuard,
+    func: fn(DisabledLocalIrqGuard, u8) -> DisabledLocalIrqGuard,
 ) {
     BOTTOM_HALF_HANDLER_L1.call_once(|| func);
 }
@@ -31,15 +31,15 @@ pub fn register_bottom_half_handler_l1(
 /// The function takes the IRQ number being processed as argument.
 ///
 /// The function may be called only once; subsequent calls take no effect.
-pub fn register_bottom_half_handler_l2(func: fn(usize)) {
+pub fn register_bottom_half_handler_l2(func: fn(u8)) {
     BOTTOM_HALF_HANDLER_L2.call_once(|| func);
 }
 
-static BOTTOM_HALF_HANDLER_L1: Once<fn(DisabledLocalIrqGuard, usize) -> DisabledLocalIrqGuard> =
+static BOTTOM_HALF_HANDLER_L1: Once<fn(DisabledLocalIrqGuard, u8) -> DisabledLocalIrqGuard> =
     Once::new();
-static BOTTOM_HALF_HANDLER_L2: Once<fn(usize)> = Once::new();
+static BOTTOM_HALF_HANDLER_L2: Once<fn(u8)> = Once::new();
 
-pub(super) fn process(irq_num: usize) {
+pub(super) fn process(irq_num: u8) {
     match InterruptLevel::current() {
         InterruptLevel::L1(_) => process_l1(irq_num),
         InterruptLevel::L2 => process_l2(irq_num),
@@ -47,7 +47,7 @@ pub(super) fn process(irq_num: usize) {
     }
 }
 
-fn process_l1(irq_num: usize) {
+fn process_l1(irq_num: u8) {
     let Some(handler) = BOTTOM_HALF_HANDLER_L1.get() else {
         return;
     };
@@ -71,7 +71,7 @@ fn process_l1(irq_num: usize) {
     drop(preempt_guard);
 }
 
-fn process_l2(irq_num: usize) {
+fn process_l2(irq_num: u8) {
     let Some(handler) = BOTTOM_HALF_HANDLER_L2.get() else {
         return;
     };
